2020dev (#42)
* Deploy to jetson in separate pop-up xterm for better error visibility

* Start of 2020 beta program

* Fix native build of hw interface

* Update install dirs for 2020 wpilib

* Updates for wpilib 2020 beta 2

* Remove a bunch of dead code from ROSIterativeRobot
This really just needs to call HAL_Init, send HAL_ObserveUserProgramStarting to show
robot code ready and periodically send HAL_ObserveUserProgram*.
Changing it so it isn't derived from the WPILib robot classes gets
rid of network tables, sendable stuff and cameraserver, none of
which we care about

* Update docker run to correct container name

* Update Jenkinsfile with correct container name

* Update rtc kernel module
Various updates for roborio install script

* Various script updates

* Fix typo in deploy rio script

* Link catkin_libs against controllers - fixes missing console bridge error

* Update roborio ros setup script - mostly python tweaks

* Add python-gnupg to rio install - makes rostopic work

* Various updates to talon write() code

* Clean up debugging code from hw interface

* Formatting

* Update defaults for talon neutral mode, remove sensors

* Update to wpilib beta3

* Updates for wpilib beta 4

* Fix ros control boilerplate cross-build

* Update setup_environment to latest ZED filename and newest navX version

* Remove Jetpack 3.x kernel modules after updating to Jetpack 4.x

* Various tweaks to roborio_install_ros script

* Some clean-up of startup Jetson scripts

* Don't build sim for anything but laptops
Updated comments in setup_environment.sh

* hwi for rio - don't need camera server libs or any of the opencv one
Remove from link step to save time and from roborio_install_ros to save
space on Rio

* Update to wpilib 2020 beta4

* Don't install jni .so to roborio

* Updated ros rio binaries

* Very basic start for robot characterization code - just an example
networktables ROS node

Set up robot_characterization to be an NT server rather than client -
client is removed from the hardware interface

robot characterization - get it working in sim

Robot characterization - correctly handle cases where no drive motors found
Move robot_characterization to rio so driver station knows where to find NT server

robot_characterization : use time since node started as time value reported to host

* Add gperftools, use tcmalloc to speed up base trajectory server

* Update navx to .400

* Various fixes for cross-builds

* Update navXTimeSync submodule commit

* Install gcc/g++ 9 on Jetson

* Launch Jetson build using terminator in deploy script rather than xterm

* Add gcc-9 update_alternatives to setup_environment.sh

* Remove swerve_point_generator and path_to_goal packages, update
base_trajectory to handle stuff previously needed from those

* Install catkin tools for building
Updated various CMakeLists.txt files to catch missed package dependencies

* Missed a build dependency

* Switch to using catkin tools for building vs. catkin_make

* Blacklist ZED example packages from native_build

* Use nt flush to push data out of robot_characterization node

* Clean up signed/unsigned compiler warning

* Rename robot interface num_pwms to be consistent with the rest of the num_* vars

* Fix update rate on robot_characterization node

* Remove unneeded dependencty from behaviors package

* Split talon_state_controller messages into separate package to help with build dependencies

* Add new catkin tools dirs to .gitignore

* Remove ros_control_boilerplate dependency from behaviors

* Fix talon_swerve controller package.xml to match CMakeLists.txt

* More cleanup of package dependencies

* Split behavior actions into a separate package to expose more parallelism in builds

* Fix dependency info for rqt_driver_station_sim

* Fix terminator call in deploy.sh

* Clean up compiler warnings

* Clean all build types (cross vs native) in clean.sh

* deploy.sh - exclude more large files not needed on the Jetson,
run terminator with -XC option

* Teleop joystick node shouldn't depend on robot visualization

* Exclude more packages from Rio & jetson build - hopefully they'll
speed things up

* Clean up common options in CMakeOpt.cmake

* Split controllers_2019 msgs into separate package to break dependency
chain where behaviors and teleop code needs to wait for the controllers
themselves to be built

* Revert back to default gcc7

* Enable color and window resizing on deploy script Jetson popup

* Remove comment from setup_environment which was breaking the apt install
line

* Add realtime_tools dependency to teleop_joystick

* Don't launch robot_visualizer

* Remove unused realtime_buffer include from teleop_joystick

* comment out gazebo install so it doesn't end up on the Jetson by mistake

* Update to non-beta container name in docker-run

* Update to 2020.1.2 WPIlib kickoff release

+++ zebROS_ws/src/rosbag_scripts/src/rename_frame_id.cpp
@@ -37,7 +37,7 @@ foreach(rosbag::MessageInstance const m, view_sick)
 rosbag::Bag write_bag;
 write_bag.open("_2018-11-03-11-08-20.bag.active", rosbag::bagmode::Write);
 
-for(int i = 0; i < laser_msg_write.size(); i++)
+for(size_t i = 0; i < laser_msg_write.size(); i++)
 {
     write_bag.write("/sick_tim/scan_modified", laser_msg_write[i].header.stamp, laser_msg_write[i]);
 }
